[리팩토링] 유저의 게임 정보 조회 기능을 User쪽에서 분리
