+++ backend/src/main/java/com/shinhan/walfi/domain/game/UserGameInfo.java
@@ -32,7 +32,4 @@ public class UserGameInfo {
     @OneToMany(mappedBy = "userGameInfo")
     private List<GameCharacter> gameCharacters = new ArrayList<>();
 
-    public void setStatus(String status) {
-        this.status = status;
-    }
 }
