+++ backend/src/main/java/com/shinhan/walfi/service/game/UserGameInfoService.java
@@ -1,0 +1,9 @@
+package com.shinhan.walfi.service.game;
+
+import com.shinhan.walfi.dto.game.UserGameInfoDto;
+
+public interface UserGameInfoService {
+
+    UserGameInfoDto getUserGameInfo(String userId);
+
+}
